requirements.txt
prettytable==3.3.0
pycryptodome==3.15.0
pyfiglet==0.8.post1
pyperclip==1.8.2
termcolor==1.1.0
colorama==0.4.5
